(rootpy) Included first function for CMS plots
Included the first function for CMS plots (called 'test'), at the
moment it makes a bar chart for the background, stacked with a bar
chart for the signal (at the moment a higgs) and pseudo data. Also
a ratio for data over background and data over signal plus background
is included.