Fixed the error band plotting for non log plots
For log plots matplotlib needs a workaround to draw error bands
correctly (basically restricting the band range to positive values)
This dosn't work for the error bands in the additional plots.
Therefore it is now checked if the plot is in log scale and only
then the positive requirement is imposed.